backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
python-dotenv==1.0.1
openai==1.59.3
pypdf==5.1.0
openpyxl==3.1.5
python-docx>=1.1.0
pymupdf>=1.24.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.100.0
apscheduler>=3.10.0
plyer>=2.1.0
watchdog>=4.0.0
duckduckgo-search>=5.0.0
httpx>=0.25.0
